+++ geo_areas/forms.py
@@ -30,152 +30,6 @@ class GeographicalAreaCreateDescriptionForm(CreateDescriptionForm):
         fields = ("described_geographicalarea", "description", "validity_start")
 
 
-class GeoAreaForm(forms.Form):
-    geo_area = forms.ModelChoiceField(
-        label="",
-        queryset=GeographicalArea.objects.all(),
-        help_text="Select a country or region.",
-        required=False,
-        widget=forms.Select(attrs={"class": "govuk-select"}),
-    )
-
-    def __init__(self, *args, **kwargs):
-        tx = kwargs.pop("transaction", None)
-        self.transaction = tx
-        super().__init__(*args, **kwargs)
-        self.fields["geo_area"].queryset = with_latest_description_string(
-            GeographicalArea.objects.exclude(
-                area_code=AreaCode.GROUP,
-                descriptions__description__isnull=True,
-            )
-            .as_at_today()
-            .approved_up_to_transaction(tx)
-            .with_latest_links("descriptions")
-            .prefetch_related("descriptions")
-            .order_by("descriptions__description"),
-            # descriptions__description" should make this implicitly distinct()
-        )
-        for field in ["geo_area"]:
-            self.fields[field].label_from_instance = lambda obj: obj.description
-
-    def clean(self, *args, **kwargs):
-        cleaned_data = super().clean()
-        return cleaned_data
-
-
-GeoAreaFormSet = formset_factory(
-    GeoAreaForm,
-    formset=FormSet,
-    min_num=1,
-    max_num=2,
-    extra=1,
-    validate_min=True,
-    validate_max=True,
-)
-
-
-class GeographicalAreaFormMixin(forms.Form):
-    """
-    Adds a geographical area selection field.
-
-    Consists of a radio group of 3 sections:
-      1. Erga Omnes (all countries), containing a subfield allowing the selection of
-         multiple geographical areas to exclude,
-      2. A group of countries, containing an autocomplete field for the primary group, and
-         another subfield for exclusions from that group,
-      3. Specific countries or regions, containing a formset with 1 autocomplete field for the country or
-         region. A maximum of 2 can be added.
-    """
-
-    class GeoAreaType(models.TextChoices):
-        ERGA_OMNES = "ERGA_OMNES", "All countries (erga omnes)"
-        GROUP = "GROUP", "A group of countries"
-        COUNTRY = "COUNTRY", "Specific countries or regions"
-
-    geo_area_type = forms.ChoiceField(choices=GeoAreaType.choices, required=False)
-    erga_omnes_exclusions = forms.ModelMultipleChoiceField(
-        queryset=GeographicalArea.objects.all(),
-        help_text="To exclude countries, enter them below.",
-        required=False,
-    )
-    geo_group = forms.ModelChoiceField(
-        queryset=GeographicalArea.objects.all(),
-        help_text="Select a country group.",
-        required=False,
-        widget=forms.Select(attrs={"class": "govuk-select"}),
-    )
-    geo_group_exclusions = forms.ModelMultipleChoiceField(
-        queryset=GeographicalArea.objects.all(),
-        help_text="Select country exclusions.",
-        required=False,
-    )
-
-    def __init__(self, *args, **kwargs):
-        tx = kwargs.pop("transaction", None)
-        self.transaction = tx
-        super().__init__(*args, **kwargs)
-
-        self.fields["geo_group"].queryset = with_latest_description_string(
-            GeographicalArea.objects.filter(
-                area_code=AreaCode.GROUP,
-            )
-            .exclude(descriptions__description__isnull=True)
-            .as_at_today()
-            .approved_up_to_transaction(tx)
-            .with_latest_links("descriptions")
-            .prefetch_related("descriptions")
-            .order_by("descriptions__description"),
-            # descriptions__description" should make this implicitly distinct()
-        )
-        # self.fields[
-        #     "geo_group_exclusions"
-        # ].queryset = GeographicalArea.objects.approved_up_to_transaction(tx)
-
-        for field in ["geo_group"]:
-            self.fields[field].label_from_instance = lambda obj: obj.description
-
-    def clean(self):
-        cleaned_data = super().clean()
-
-        geo_area_type = cleaned_data.pop("geo_area_type", None)
-        erga_omnes_exclusions = cleaned_data.pop("erga_omnes_exclusions", None)
-        geo_group = cleaned_data.pop("geo_group", None)
-        geo_group_exclusions = cleaned_data.pop("geo_group_exclusions", None)
-        geo_area = cleaned_data.pop("geo_area", None)
-
-        if geo_area_type == self.GeoAreaType.ERGA_OMNES:
-            cleaned_data["geographical_area"] = (
-                GeographicalArea.objects.approved_up_to_transaction(self.transaction)
-                .erga_omnes()
-                .get()
-            )
-            cleaned_data["geo_area_exclusions"] = erga_omnes_exclusions
-
-        self.fields["geo_area_type"].initial = geo_area_type
-
-        if geo_area_type == self.GeoAreaType.GROUP:
-            if not geo_group:
-                raise ValidationError({"geo_group": "A country group is required."})
-            cleaned_data["geographical_area"] = geo_group
-            cleaned_data["geo_area_exclusions"] = geo_group_exclusions
-
-        if geo_area_type == self.GeoAreaType.COUNTRY:
-            if not geo_area:
-                raise ValidationError("One or more countries or regions is required.")
-            cleaned_data["geographical_area"] = geo_area
-
-        self.fields["geo_group"].initial = geo_group.pk if geo_group else None
-
-        if not cleaned_data["geographical_area"]:
-            raise ValidationError("A Geographical area must be selected")
-
-        return cleaned_data
-
-
-class GeographicalAreaSelect(Field):
-    template = "components/geographical_area_select/template.jinja"
-
-
 GeographicalAreaDeleteForm = delete_form_for(GeographicalArea)
